DB und Zeitaufräumen (#371)
* [DB] fix trigger time format

* fix min/max values

* clean stock exchange news trigger

* invisible start trigger
* self-triggering followup news
* my grammar currently requires quality and price

* [Misc] remove time parameter from MakeRealTime

A real date already corresponds to particular game time, so remove time
parameters from time type 4 and 5 (see #364 and #366).

* [Misc] fix one-off in time calculation

When determinining the time for a particular game day from a user's time
definition, one day must be substracted. For the user day 1 is the first
day, internally it is day 0.